requirements.txt
Flask
python-dotenv
resend
